Create Number Guessing game.cpp
A C++ program (Number Guessing game) that generates a random number and asks the user to guess it. Provide feedback on whether the guess is too high or too low until the user guesses the correct number.
The program will provide the user with a a more helpful hint by indicating the range of possible numbers after the first incorrect guess.